(index.lisp) add-indexes: SBCL Workaround
Although modifying &REST lambda variables is not strictly forbidden by
the CLHS, it is discouraged[1]. When compiling under safety values less
than 3 SBCL assumes the &REST parameter to be a list, so calls to elt are
specialized to the list variation. Rebind the rest parameter to itself
as a workaround.

[1]: "There is a particular problem with &REST lambda variables, which
are always bound to a value of type LIST."
cf. http://www.lispworks.com/documentation/HyperSpec/Issues/iss178_w.htm